fix the BlockLength should be int32 not uint16 in ErrorInfo struct

+++ internal/info/info_error.go
@@ -15,7 +15,7 @@ type ErrorItem struct {
 
 type ErrorInfo struct {
 	HeaderBlockNumber 	uint8
-	BlockLength 		uint16
+	BlockLength 		int32
 	ErrorNumber			uint16
 	ErrorItems			[]ErrorItem
 	Spare				[40]byte
